adds namespace to services
